+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
   the previous UI widget drawn.
 - Added `PixState::set_column_offset` and `PixState::reset_column_offset` to
   allow controlling the x-offset position when rendering UI elements.
+- Added `ThemeBuilder` to prelude.
 
 ### Changed